Test File
This can be used for testing this program. The values are loaded with three columns. the first column indicates if the person is either male female or a zombie. The second column indicates when that particular person arrived to the bathroom. The third column indicates how long that person was in the bathroom.